msm: vidc: Add VP9 decoder support on msm8956/76 version v1
Video hardware on msm8956/76 version v1 chipset supports
VP9 decoder with a hardware limitation which requires
more voltage for a given video clock rate. So read
hardware version from efuse register and do necessary
voltage scaling by defining new VP9 decoder clock vs
voltage table in dtsi and use it when VP9 decoder
session is running.

+++ drivers/media/platform/msm/vidc/venus_hfi.h
@@ -229,7 +229,7 @@ struct venus_hfi_device {
 	struct list_head sess_head;
 	u32 intr_status;
 	u32 device_id;
-	u32 clk_freq;
+	unsigned long clk_freq;
 	u32 last_packet_type;
 	unsigned long clk_bitrate;
 	unsigned long scaled_rate;
